move argument parsing to args.rs
