+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod app;
+pub mod args;
 pub mod cli;
 pub mod execution;
 pub mod telemetry;
